Blog search by filter
Added in the ability in the blog get method, now we can filter what blogs a person will look at. Example, by game, by title, and they can be ordered by most recent, oldest, and by likes. Also made it so to delete/update a blog or comment, the user must be the same user who created the blog/comment